Remove `collection` module
